refactor: centralize model manager
