+++ agent_forge/models/__init__.py
@@ -1,0 +1,5 @@
+"""Model utilities for Agent Forge."""
+
+from .manager import SimpleModelManager
+
+__all__ = ["SimpleModelManager"]
