+++ start_agent_forge.py
@@ -6,11 +6,8 @@ Downloads 3 seed models to D: drive and starts EvoMerge evolution process.
 import asyncio
 import json
 import logging
-import shutil
-import subprocess
-import sys
 from datetime import datetime
-from pathlib import Path
+from agent_forge.models import SimpleModelManager
 
 # Configure logging
 logging.basicConfig(
@@ -36,172 +33,6 @@ SEED_MODELS = [
         "description": "1.3B parameter efficient Python model",
     },
 ]
-
-
-class SimpleModelManager:
-    """Simplified model manager for production pipeline."""
-
-    def __init__(self, base_dir="D:/AIVillage/models", max_models=8):
-        self.base_dir = Path(base_dir)
-        self.max_models = max_models
-        self.models = {}
-
-        # Create base directory
-        self.base_dir.mkdir(parents=True, exist_ok=True)
-        logger.info(f"Model storage: {self.base_dir}")
-
-        # Check available space
-        self._check_storage_space()
-
-        # Load existing models
-        self._load_existing_models()
-
-    def _check_storage_space(self):
-        """Check available storage space."""
-        try:
-            free_bytes = shutil.disk_usage(self.base_dir).free
-            free_gb = free_bytes / (1024**3)
-            logger.info(f"Available storage: {free_gb:.1f} GB")
-
-            if free_gb < 15:
-                logger.warning("Low storage space - models may not download properly")
-
-        except Exception as e:
-            logger.warning(f"Could not check storage: {e}")
-
-    def _load_existing_models(self):
-        """Load information about existing models."""
-        try:
-            for model_dir in self.base_dir.iterdir():
-                if model_dir.is_dir():
-                    # Extract model ID from directory name
-                    model_id = model_dir.name.replace("_", "/")
-                    if model_id.count("/") == 1:  # Valid model ID format
-                        size_gb = self._get_directory_size(model_dir) / (1024**3)
-                        self.models[model_id] = {
-                            "path": str(model_dir),
-                            "size_gb": size_gb,
-                            "downloaded": True,
-                        }
-                        logger.info(
-                            f"Found existing model: {model_id} ({size_gb:.1f} GB)"
-                        )
-
-            logger.info(f"Loaded {len(self.models)} existing models")
-
-        except Exception as e:
-            logger.error(f"Error loading existing models: {e}")
-
-    def _get_directory_size(self, path):
-        """Get total size of directory in bytes."""
-        total = 0
-        try:
-            for file_path in path.rglob("*"):
-                if file_path.is_file():
-                    total += file_path.stat().st_size
-        except Exception:
-            pass
-        return total
-
-    async def download_model(self, model_spec):
-        """Download a model using huggingface-cli."""
-        model_id = model_spec["model_id"]
-
-        # Check if already downloaded
-        if model_id in self.models:
-            logger.info(f"Model {model_id} already downloaded")
-            return True
-
-        # Create safe directory name
-        safe_name = model_id.replace("/", "_").replace(":", "_")
-        model_dir = self.base_dir / safe_name
-
-        logger.info(f"Downloading {model_id} to {model_dir}")
-
-        try:
-            # Use huggingface-cli for downloading
-            cmd = [
-                "huggingface-cli",
-                "download",
-                model_id,
-                "--local-dir",
-                str(model_dir),
-                "--local-dir-use-symlinks",
-                "False",
-            ]
-
-            logger.info(f"Running: {' '.join(cmd)}")
-
-            # Run download command
-            process = subprocess.Popen(
-                cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True
-            )
-
-            stdout, stderr = process.communicate()
-
-            if process.returncode == 0:
-                # Calculate size
-                size_gb = self._get_directory_size(model_dir) / (1024**3)
-
-                # Update registry
-                self.models[model_id] = {
-                    "path": str(model_dir),
-                    "size_gb": size_gb,
-                    "downloaded": True,
-                    "purpose": model_spec["purpose"],
-                }
-
-                logger.info(f"Successfully downloaded {model_id} ({size_gb:.1f} GB)")
-                return True
-            else:
-                logger.error(f"Download failed for {model_id}")
-                logger.error(f"Error: {stderr}")
-                return False
-
-        except FileNotFoundError:
-            logger.error("huggingface-cli not found - installing huggingface_hub")
-
-            # Try to install huggingface_hub
-            try:
-                subprocess.run(
-                    [sys.executable, "-m", "pip", "install", "huggingface_hub"],
-                    check=True,
-                )
-                logger.info("Installed huggingface_hub - please restart the script")
-                return False
-            except subprocess.CalledProcessError:
-                logger.error("Failed to install huggingface_hub")
-                return False
-
-        except Exception as e:
-            logger.error(f"Error downloading {model_id}: {e}")
-            return False
-
-    def cleanup_old_models(self):
-        """Remove old models if we exceed max_models limit."""
-        if len(self.models) <= self.max_models:
-            return
-
-        logger.info(f"Cleaning up models: {len(self.models)} > {self.max_models}")
-
-        # Sort by size (remove largest first to save space quickly)
-        models_by_size = sorted(
-            self.models.items(), key=lambda x: x[1].get("size_gb", 0), reverse=True
-        )
-
-        to_remove = models_by_size[self.max_models :]
-
-        for model_id, model_info in to_remove:
-            try:
-                model_path = Path(model_info["path"])
-                if model_path.exists():
-                    shutil.rmtree(model_path)
-                    logger.info(f"Removed {model_id}")
-
-                del self.models[model_id]
-
-            except Exception as e:
-                logger.error(f"Failed to remove {model_id}: {e}")
 
 
 class SimpleEvoMerge:
